--- src/main/java/create_task_use_case/CreateTaskInputBoundary.java
@@ -1,11 +1,0 @@
-/**
- * Create Task Input Boundary Interface.
- * Author: Fardin Faruk
- * Modified By: N/A
- * Created: Nov 19, 2022
- * Last Modified: N/A
- */package create_task_use_case;
-
-public interface CreateTaskInputBoundary {
-    CreateTaskOutputData create(CreateTaskInputData taskData);
-}
