+++ src/test/java/create_task_use_case/CreateTaskOutputBoundary.java
@@ -1,12 +1,10 @@
-/**
- * Create Task Output Boundary Interface.
- * Author: Fardin Faruk
- * Modified By: Fardin Faruk
- * Created: Nov 19, 2022
- * Last Modified: Nov 27, 2022
- */
 package create_task_use_case;
 
+/**
+ * Task Creation Output Boundary Interface. Provides the prepareSuccessView and
+ * prepareFailView methods which are implemented by ModifyTaskPresenter.
+ * @author Fardin Faruk
+ */
 public interface CreateTaskOutputBoundary {
     CreateTaskOutputData prepareSuccessView(CreateTaskOutputData response);
