+++ src/test/java/create_task_use_case/CreateTaskOutputData.java
@@ -1,12 +1,12 @@
-/**
- * Create Task Output Data Class.
- * Author: Fardin Faruk
- * Modified By: Fardin Faruk
- * Created: Nov 19, 2022
- * Last Modified: Nov 27, 2022
- */
 package create_task_use_case;
 
+/**
+ * Task Creation Output Data Class. Instances of this class are returned upon the
+ * execution of the application logic for this use case. The errorMessage and
+ * isSuccessfullyModified instance attributes, in particular, represent whether the use
+ * case logic was successfully executed.
+ * @author Fardin Faruk
+ */
 public class CreateTaskOutputData {
 
     String title;
