+++ src/test/java/create_task_use_case/CreateTaskOutputDataTest.java
@@ -2,7 +2,6 @@ package create_task_use_case;
 
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class CreateTaskOutputDataTest {
